Migrate `serializeAttribute()` from privacy.ts to serialize.ts
