+++ packages/rum/src/domain/record/privacy.ts
@@ -11,7 +11,6 @@ import {
   PRIVACY_CLASS_HIDDEN,
   FORM_PRIVATE_TAG_NAMES,
   CENSORED_STRING_MARK,
-  CENSORED_IMG_MARK,
   // Deprecated (now aliased) below
   PRIVACY_CLASS_INPUT_IGNORED,
   PRIVACY_CLASS_INPUT_MASKED,
@@ -21,7 +20,7 @@ import {
 
 export const MAX_ATTRIBUTE_VALUE_CHAR_LENGTH = 100_000
 
-import { makeStylesheetUrlsAbsolute, makeSrcsetUrlsAbsolute, makeUrlAbsolute } from './serializationUtils'
+import { makeStylesheetUrlsAbsolute } from './serializationUtils'
 
 import { shouldIgnoreElement } from './serialize'
 
@@ -150,66 +149,6 @@ export function shouldMaskNode(node: Node, privacyLevel: NodePrivacyLevel) {
       return isTextNode(node) ? isFormElement(node.parentNode) : isFormElement(node)
     default:
       return false
-  }
-}
-
-export function serializeAttribute(
-  element: Element,
-  nodePrivacyLevel: NodePrivacyLevel,
-  attributeName: string
-): string | number | boolean | null {
-  if (nodePrivacyLevel === NodePrivacyLevel.HIDDEN) {
-    // dup condition for direct access case
-    return null
-  }
-  const attributeValue = element.getAttribute(attributeName)
-  if (nodePrivacyLevel === NodePrivacyLevel.MASK) {
-    const tagName = element.tagName
-
-    switch (attributeName) {
-      // Mask Attribute text content
-      case 'title':
-      case 'alt':
-        return CENSORED_STRING_MARK
-    }
-    // mask image URLs
-    if (tagName === 'IMG' || tagName === 'SOURCE') {
-      if (attributeName === 'src' || attributeName === 'srcset') {
-        return CENSORED_IMG_MARK
-      }
-    }
-    // mask <a> URLs
-    if (tagName === 'A' && attributeName === 'href') {
-      return CENSORED_STRING_MARK
-    }
-    // mask data-* attributes
-    if (attributeValue && attributeName.indexOf('data-') === 0 && attributeName !== PRIVACY_ATTR_NAME) {
-      // Exception: it's safe to reveal the `${PRIVACY_ATTR_NAME}` attr
-      return CENSORED_STRING_MARK
-    }
-  }
-
-  if (!attributeValue || typeof attributeValue !== 'string') {
-    return attributeValue
-  }
-
-  // Minimum Fix for customer.
-  if (attributeValue.length > MAX_ATTRIBUTE_VALUE_CHAR_LENGTH && attributeValue.slice(0, 5) === 'data:') {
-    return 'data:truncated'
-  }
-
-  // Rebuild absolute URLs from relative (without using <base> tag)
-  const doc = element.ownerDocument
-  switch (attributeName) {
-    case 'src':
-    case 'href':
-      return makeUrlAbsolute(attributeValue, doc.location?.href)
-    case 'srcset':
-      return makeSrcsetUrlsAbsolute(attributeValue, doc.location?.href)
-    case 'style':
-      return makeStylesheetUrlsAbsolute(attributeValue, doc.location?.href)
-    default:
-      return attributeValue
   }
 }
